Rich Html previewer updates
Prior code did not handle i18n correctly and would pop up the
confirmation dialog on the main file page when launching the previewer
in a separate window, causing the previewer to hang/fail. The return
option also created a second tab with the dataset. This PR replaces the
popup with an inline warning and the option to Display With/Without the
Rich Content, and, if the data doesn't contain scripts directly or a
<head> element, just displays it as the HTML previewer would (making
this a possible replacement for the HTML Previewer rather than an
optional second previewer.) Also ran some auto-translations into es and
fr.