Use stock script
For pre-Dataverse 6.4 (which doesn't recognize the *.glb files, one can
edit the Accept header in the voyager-explorer.min.js script to accept *
in addition to text/plain (there is only one place this happens in their
current script so easy to find). 6.4, in addition to adding the
mimetype, changes the file download api to @Produce */* (making
text/plain an OK request rather than a 406 error).